Move highlighting code to metabase/ui/syntax
