+++ frontend/src/metabase/ui/syntax/highlight.module.css
