+++ frontend/src/metabase/ui/syntax/highlight.ts
