+++ frontend/src/metabase/ui/syntax/index.ts
@@ -1,0 +1,1 @@
+export * from "./highlight";
